refactor: improve `proof-of-sql-parser` - put `test_utility.rs` in parser behind a feature flag and allow it to be used by the other crate - prevent `posql_time` from being unnecessarily convoluted
