+++ crates/proof-of-sql-parser/src/sql.lalrpop
@@ -2,7 +2,7 @@ use crate::intermediate_ast;
 use crate::select_statement;
 use crate::identifier;
 use lalrpop_util::ParseError::User;
-use crate::{intermediate_decimal::IntermediateDecimal, posql_time::timestamp::PoSQLTimestamp};
+use crate::{intermediate_decimal::IntermediateDecimal, posql_time::PoSQLTimestamp};
 
 grammar;
