+++ crates/proof-of-sql/src/base/commitment/column_commitment_metadata.rs
@@ -169,10 +169,7 @@ mod tests {
         commitment::column_bounds::Bounds, database::OwnedColumn, math::decimal::Precision,
         scalar::Curve25519Scalar,
     };
-    use proof_of_sql_parser::posql_time::{
-        timezone::{self, PoSQLTimeZone},
-        unit::PoSQLTimeUnit,
-    };
+    use proof_of_sql_parser::posql_time::{PoSQLTimeUnit, PoSQLTimeZone};
 
     #[test]
     fn we_can_construct_metadata() {
@@ -374,7 +371,7 @@ mod tests {
         let timestamp_column: OwnedColumn<Curve25519Scalar> =
             OwnedColumn::<Curve25519Scalar>::TimestampTZ(
                 PoSQLTimeUnit::Second,
-                timezone::PoSQLTimeZone::Utc,
+                PoSQLTimeZone::Utc,
                 [1i64, 2, 3, 4, 5].to_vec(),
             );
         let committable_timestamp_column = CommittableColumn::from(&timestamp_column);
@@ -536,7 +533,7 @@ mod tests {
             1_625_072_400,
             1_625_065_000,
         ];
-        let timezone = timezone::PoSQLTimeZone::Utc;
+        let timezone = PoSQLTimeZone::Utc;
         let timeunit = PoSQLTimeUnit::Second;
         let timestamp_column_a = CommittableColumn::TimestampTZ(timeunit, timezone, &times[..2]);
         let timestamp_metadata_a = ColumnCommitmentMetadata::from_column(&timestamp_column_a);
@@ -562,7 +559,7 @@ mod tests {
             1_625_072_400,
             1_625_065_000,
         ];
-        let timezone = timezone::PoSQLTimeZone::Utc;
+        let timezone = PoSQLTimeZone::Utc;
         let timeunit = PoSQLTimeUnit::Second;
 
         let timestamp_column_a = CommittableColumn::TimestampTZ(timeunit, timezone, &times[..2]);
@@ -859,18 +856,12 @@ mod tests {
             .is_err());
 
         let timestamp_tz_metadata_a = ColumnCommitmentMetadata {
-            column_type: ColumnType::TimestampTZ(
-                PoSQLTimeUnit::Second,
-                timezone::PoSQLTimeZone::Utc,
-            ),
+            column_type: ColumnType::TimestampTZ(PoSQLTimeUnit::Second, PoSQLTimeZone::Utc),
             bounds: ColumnBounds::TimestampTZ(Bounds::Empty),
         };
 
         let timestamp_tz_metadata_b = ColumnCommitmentMetadata {
-            column_type: ColumnType::TimestampTZ(
-                PoSQLTimeUnit::Millisecond,
-                timezone::PoSQLTimeZone::Utc,
-            ),
+            column_type: ColumnType::TimestampTZ(PoSQLTimeUnit::Millisecond, PoSQLTimeZone::Utc),
             bounds: ColumnBounds::TimestampTZ(Bounds::Empty),
         };
