+++ crates/proof-of-sql/src/base/database/test_accessor_utility.rs
@@ -8,7 +8,7 @@ use arrow::{
     datatypes::{i256, DataType, Field, Schema},
     record_batch::RecordBatch,
 };
-use proof_of_sql_parser::posql_time::unit::PoSQLTimeUnit;
+use proof_of_sql_parser::posql_time::PoSQLTimeUnit;
 use rand::{
     distributions::{Distribution, Uniform},
     rngs::StdRng,
